requirements.txt
sounddevice
soundfile
Flask
newsapi-python
gunicorn
firebase-admin
gradio
google-generativeai
keras
librosa
numpy
tensorflow
pillow
